refactored names.components, deleted couple unnecessary pieces of code, fixed couple grammatical mistakes0/
