+++ src/test/java/pages/сomponents/RegistrationModal.java
@@ -14,7 +14,7 @@ public class RegistrationModal {
 
     public void verifyResult(String key, String value) {
         $(".table-responsive").$(byText(key)).parent()
-                .shouldHave(text(value)); //done
+                .shouldHave(text(value));
     }
 }
